refactor(concurrency): split concurrency.py into concurrency/ package
Split 239-LOC module into:
- concurrency/protocols.py: DistributedLock, DeduplicationStore, ResultStore, TaskQueue, QueueMessage
- concurrency/memory.py: MemoryLock, MemoryDedup, MemoryResultStore, MemoryTaskQueue
- concurrency/__init__.py: re-exports all public symbols

Existing imports  package __init__ re-exports everything.unchanged
